pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "prophet>=1.1.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "statsmodels>=0.14.4",
    "streamlit>=1.46.0",
    "xgboost>=3.0.2",
]
